+++ src/components/ItaliaTheme/Blocks/CountDown/Edit.jsx
@@ -39,9 +39,13 @@ const Edit = (props) => {
               className="background-image"
               style={{
                 backgroundImage: `url(${
-                  data.background[0]?.image
+                  data.background[0]?.image_field
                     ? flattenToAppURL(
-                        data.background[0]['@id'] + '/@@images/image',
+                        data.background[0]['@id'] +
+                          '/' +
+                          data.background[0].image_scales?.[
+                            data.background[0].image_field
+                          ][0].download,
                       )
                     : addAppURL(data.background[0]?.['@id'])
                 })`,