+++ src/components/ItaliaTheme/Blocks/Listing/InEvidenceTemplate.jsx
@@ -1,3 +1,6 @@
+/*
+ * In evidenza
+ */
 import React from 'react';
 import PropTypes from 'prop-types';
 import {
@@ -83,7 +86,10 @@ const InEvidenceTemplate = (props) => {
             const listingText = show_description ? (
               <ListingText item={item} />
             ) : null;
-            const image = ListingImage({ item });
+            const image = ListingImage({
+              item,
+              sizes: '(max-width:320px) 200px, 300px',
+            });
             const category = getCategory(item, show_type, show_section, props);
             const topics = show_topics ? item.tassonomia_argomenti : null;
 
@@ -98,12 +104,7 @@ const InEvidenceTemplate = (props) => {
                 key={index}
               />
             ) : (
-              <Card
-                key={index}
-                className={cx('listing-item card-bg', {
-                  'card-img': index === 0 && image,
-                })}
-              >
+              <Card key={index} className={cx('listing-item card-bg')}>
                 {index === 0 && image && (
                   <div className="img-responsive-wrapper">
                     <div className="img-responsive">